+++ internal/table/params.go
@@ -5,14 +5,17 @@ import (
 	"fmt"
 	"sort"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
 )
 
+var ErrNameRequired = xerrors.Wrap(fmt.Errorf("ydb: only named parameters are supported"))
+
 // GenerateDeclareSection generates DECLARE section text in YQL query by params
 //
 // Warning: This is an experimental feature and could change at any time
-func GenerateDeclareSection(params *table.QueryParameters) string {
+func GenerateDeclareSection(params *table.QueryParameters) (string, error) {
 	var (
 		buf      bytes.Buffer
 		names    []string
@@ -28,7 +31,10 @@ func GenerateDeclareSection(params *table.QueryParameters) string {
 	})
 	sort.Strings(names)
 	for _, name := range names {
+		if name == "" {
+			return "", xerrors.WithStackTrace(ErrNameRequired)
+		}
 		buf.WriteString(declares[name])
 	}
-	return buf.String()
+	return buf.String(), nil
 }