+++ sugar/params.go
@@ -1,3 +1,6 @@
+//go:build !go1.18
+// +build !go1.18
+
 package sugar
 
 import (
@@ -8,6 +11,6 @@ import (
 // GenerateDeclareSection generates DECLARE section text in YQL query by params
 //
 // Warning: This is an experimental feature and could change at any time
-func GenerateDeclareSection(params *table.QueryParameters) string {
+func GenerateDeclareSection(params *table.QueryParameters) (string, error) {
 	return internal.GenerateDeclareSection(params)
 }