+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Changed output of `sugar.GenerateDeclareSection` (added error as second result)
+* Specified `sugar.GenerateDeclareSection` for `go1.18` (supports input types `*table.QueryParameters` `[]table.ParameterOption` or `[]sql.NamedArg`)
+* Added `database/sql` example and docs (WIP)
+
 ## v3.36.1
 * Fixed `xsql.Rows` error checking